refactor: ProductπουργSearchMv 삭제 – 검색을 tsvector 기반으로 통합
